Update coord_to_mx, swap product ord for RST ops
The previous version of `coord_to_mx` by default assumed the output
matrix will be of 1x3 dimension. This commit adds a row and a column
parameter to the function so that the user can define the dimension of
the output matrix.

The other update this commit includes is an update to the order of the
matrix multiplication. The product terms are swapped, because of the
transposing in the coordinate dimensions.