Fix up vector edge case
When the up vector coincided with the forward vector, a new hard coded
vector was introduced, but the image produced was flipped incorrectly.
This commit fixes the issue, and when the camera looks down, (y_vec=-1)
the up vector is z_vec=1. When the camera looks upwards (y_vec=1) the
up vector is z_vec=-1.